requirements.txt
numpy
PyOpenGL
glfw
imgui